+++ web/pdf_viewer.component.js
@@ -14,23 +14,14 @@
  */
 
 import {
-  AnnotationLayerBuilder,
   DefaultAnnotationLayerFactory,
-} from "./annotation_layer_builder.js";
-import {
   DefaultStructTreeLayerFactory,
-  StructTreeLayerBuilder,
-} from "./struct_tree_layer_builder.js";
-import {
   DefaultTextLayerFactory,
-  TextLayerBuilder,
-} from "./text_layer_builder.js";
-import {
   DefaultXfaLayerFactory,
-  XfaLayerBuilder,
-} from "./xfa_layer_builder.js";
+} from "./default_factory.js";
 import { PDFLinkService, SimpleLinkService } from "./pdf_link_service.js";
 import { PDFSinglePageViewer, PDFViewer } from "./pdf_viewer.js";
+import { AnnotationLayerBuilder } from "./annotation_layer_builder.js";
 import { DownloadManager } from "./download_manager.js";
 import { EventBus } from "./event_utils.js";
 import { GenericL10n } from "./genericl10n.js";
@@ -40,6 +31,9 @@ import { PDFHistory } from "./pdf_history.js";
 import { PDFPageView } from "./pdf_page_view.js";
 import { PDFScriptingManager } from "./pdf_scripting_manager.js";
 import { ProgressBar } from "./ui_utils.js";
+import { StructTreeLayerBuilder } from "./struct_tree_layer_builder.js";
+import { TextLayerBuilder } from "./text_layer_builder.js";
+import { XfaLayerBuilder } from "./xfa_layer_builder.js";
 
 // eslint-disable-next-line no-unused-vars
 const pdfjsVersion = PDFJSDev.eval("BUNDLE_VERSION");
